Altered a lot of code to adhere to the lecture on 10/26, see below:
I've simplified the move logic for the Pawn, and for the
lateral(horizontal and vertical) movements and the diagonal movements.
It's more efficient and should allow easier modifications, especially
for adding en passant to the Pawn. On top of that I also altered the
layout of the the entire program. Within each of the Pieces there is now
a new overridden method that checks their moves, this is if they have a
unique move. I've added a Movement class for all of the basic movement
checking, lateral and diagonal. This was done to make the classes more
uniform in size and make the "lazy" classes, the pieces, take more of
the work.